requirements.txt
Django>=2.0
django-bootstrap3>=9.1.0
django-leaflet==0.19.0
googlemaps>=2.5.1
gpxpy>=1.2.0
